Add a 1.8V regulator, jumpers for the UART, and LEDs
This commit adds a 1.8V regulator to the board design. Originally it had
been based on a simple diode drop, which may still be the better option,
but it is a good idea to have the option of soldering down a regulator
at this prototyping stage. Adding the pads doesn't add to the cost of
the board.

Also add jumper pads so that either UART0 or UART1 can be selected, and
add LEDs to show the status of the voltage rails. Reset and power lines
are wired up to GPIO0 & GPIO1, but those are also the lines for RX & TX
activity. I've added LEDs to those lines to show when either go low, but
also added jumper pads so it can be disconnected from reset/pwr and be
used for activity display.

Finally, drop the gpio_wkup connection. It would be useful to have that,
but not just yet. I will consider adding that back in the future.

Signed-off-by: Grant Likely <grant.likely@linaro.org>