Move OPL part numbers to user field 1
The BOM generation tool doesn't know how to sort user fields by name, so
it puts things in the wrong columns if parts have user data in a
different order. Move the OPL part number to user field 1 always so that
it can be easily used for automatic BOM generation.

Signed-off-by: Grant Likely <grant.likely@linaro.org>