Get rid of rescued components from kicad uprev
The uprev to the latest KiCad caused some components to conflict with
what is in the standard library. KiCad created 'rescued' components to
keep the schematic readable, but it isn't really ideal because the
standard footprints don't necessarily match. Fix up the schematic to use
standard components and update the footprint assignments to match.

Changed components: GND, +1.8V, +3.3V, R, Led_Small

Signed-off-by: Grant Likely <grant.likely@linaro.org>