Fix encoding of FTProg FT230X template file.
The FT230X template file was stored in git as a UTF-8 encoded XML file.
Unfortunately, FTDI's FTProg utility doesn't play nice with that
encoding. It will only emit and read UTF-16le encoded files. Change the
file to use the required UTF-16le encoding.

This isn't particularly friendly for Linux editor tools, but we have to
deal with what the tool will accept, and look forward to the day when
someone writes a Linux tool for flashing the FT-230X.

Signed-off-by: Grant Likely <grant.likely@linaro.org>