Add direct GPIO control
Sometimes it is useful to control the CBUS pins directly. Add a 'gpio'
command for directly setting the GPIO. Unfortunately, I do not know how
to read from the FTDI chip if a pin is set as input or output, so
setting a single GPIO will reset all the others back to being inputs.

Signed-off-by: Grant Likely <grant.likely@linaro.org>