96boardsctl: Fix build for libftdi >= 1.0
Currently CMakeLists.txt searches only for libftdi but on platforms
where libftdi may be parallel installed this is insufficient. Fix
the package detection and ensure we apply the detection results to
the default build arguments.

Also removes the check for libusb; 96boardsctl.c does not directly use
this library so we can rely on pkg-config to get the link line right.

Signed-off-by: Daniel Thompson <daniel.thompson@linaro.org>